requirements.txt
# Core Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# LangChain & LangGraph (Orchestration)
langchain==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5
langgraph==0.1.5
langchain-chroma==0.1.1

# Vector Database
chromadb==0.5.0

# PDF Processing
PyMuPDF==1.24.4

# Web Scraping (Ingestion Agent)
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.2.2

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
tiktoken==0.7.0
jinja2==3.1.4
aiofiles==23.2.1

# difflib is part of Python stdlib — no install needed

# OpenAI (default LLM)
openai==1.30.1

# Optional: Ollama (fallback LLM)
ollama==0.1.9
